jogorcu_kenesh_orm/requirements.txt
# parsing
beautifulsoup4
requests
lxml
# work with db
# work with postgres
psycopg2-binary
# orm
peewee
# for .env
python-decouple